fix(renderer): widen batch indices u16 -> u32 to stop silent overflow (RC5)
All batched 2D content (sprites, UI, text, spine, tilemap, particles) streams
into a single shared Batch vertex/index buffer. Indices were 16-bit, so once a
frame accumulated more than 65,535 batched vertices every `baseVertex` cast
truncated and the GPU read garbage geometry — a silent corruption that triggers
on any moderately busy scene (the audit's #1 latent renderer bug).

- TransientBufferPool: index_staging std::vector<u16> -> u32; writeIndices /
  appendIndices take const u32*; EBO sizing uses sizeof(u32). The const u32*
  signature makes every index-producing call site a compile error until widened.
- Draw calls that read the pool EBO now use GfxDataType::UnsignedInt with a
  sizeof(u32) element stride: DrawList (the merged batch draw) and
  ShapePlugin::customDraw.
- Widen every index array / base-vertex cast / index vector / quad-index
  constant to u32 across all batched plugins and RenderFrameSubmit. Non-index
  u16 (tile ids, text lengths, UTF-8 decode) left unchanged.

Verified: emcc 5.0.0 web build compiles clean and links esengine-core.wasm. The
u32 signature surfaced the one site outside plugins/ (RenderFrameSubmit) that a
manual sweep missed. SpinePlugin is excluded from the spine-off build and was
inspection-verified (identical mechanical change).

+++ src/esengine/renderer/TransientBufferPool.hpp
@@ -41,10 +41,10 @@ public:
     u32 allocIndices(LayoutId layout, u32 count);
 
     void writeVertices(LayoutId layout, u32 byteOffset, const void* data, u32 byteSize);
-    void writeIndices(LayoutId layout, u32 indexOffset, const u16* data, u32 count);
+    void writeIndices(LayoutId layout, u32 indexOffset, const u32* data, u32 count);
 
     u32 appendVertices(LayoutId layout, const void* data, u32 byteSize);
-    u32 appendIndices(LayoutId layout, const u16* data, u32 count);
+    u32 appendIndices(LayoutId layout, const u32* data, u32 count);
 
     /** Upload every non-empty stream's staging to its VBO/EBO. */
     void upload();
@@ -66,7 +66,7 @@ private:
         u32 ebo = 0;
         u32 vao = 0;
         std::vector<u8> vertex_staging;
-        std::vector<u16> index_staging;
+        std::vector<u32> index_staging;  // 32-bit indices: a single Batch stream can exceed 65535 vertices
         u32 vertex_write_pos = 0;
         u32 index_write_pos = 0;
         u32 vbo_capacity = 0;
