+++ src/esengine/renderer/plugins/SpritePlugin.cpp
@@ -152,9 +152,9 @@ void SpritePlugin::emitQuad(
     u32 vOff = buffers.appendVertices(LayoutId::Batch, verts, sizeof(verts));
     u32 baseVertex = vOff / sizeof(BatchVertex);
 
-    u16 indices[6];
+    u32 indices[6];
     for (u32 i = 0; i < 6; ++i) {
-        indices[i] = static_cast<u16>(baseVertex + QUAD_INDICES[i]);
+        indices[i] = static_cast<u32>(baseVertex + QUAD_INDICES[i]);
     }
     u32 iOff = buffers.appendIndices(LayoutId::Batch, indices, 6);
 
@@ -246,9 +246,9 @@ void SpritePlugin::emitNineSlice(
             u32 vOff = buffers.appendVertices(LayoutId::Batch, verts, sizeof(verts));
             u32 baseVert = vOff / sizeof(BatchVertex);
 
-            u16 indices[6];
+            u32 indices[6];
             for (u32 i = 0; i < 6; ++i) {
-                indices[i] = static_cast<u16>(baseVert + QUAD_INDICES[i]);
+                indices[i] = static_cast<u32>(baseVert + QUAD_INDICES[i]);
             }
             u32 iOff = buffers.appendIndices(LayoutId::Batch, indices, 6);
 
@@ -362,9 +362,9 @@ void SpritePlugin::emitTiledQuads(
             u32 vOff = buffers.appendVertices(LayoutId::Batch, verts, sizeof(verts));
             u32 baseVert = vOff / sizeof(BatchVertex);
 
-            u16 indices[6];
+            u32 indices[6];
             for (u32 i = 0; i < 6; ++i) {
-                indices[i] = static_cast<u16>(baseVert + QUAD_INDICES[i]);
+                indices[i] = static_cast<u32>(baseVert + QUAD_INDICES[i]);
             }
             u32 iOff = buffers.appendIndices(LayoutId::Batch, indices, 6);
